modern: Handle #foo forms better.
The point of a #foo form is to "turn off" the form. Thus, we don't actually need
to handle "#" until the end of the parse. This makes things like ``GENERIC#:`` work
without extra code since the lexer no longer looks for "#".

Also clean up lex-factor-top and lex-factor-nested to reduce duplication.
There are only two cases that matter when it's "top" or "nested", which refers to
whether we are in a ``FOO:`` form or not. Another ``FOO:`` or a ``<SECTION`` can
happen inside a ``FOO:``, causing us to rewind the lex and start lexing the new form.