vocabs.loader: Delete errors after using list is fixed with restarts.
If a vocabulary needs restarts to fix its using list and that
vocabulary is used from another vocabulary that then recompiles,
the restarts were left in the `errorlist` hash even if they were
fixed. This caused that error to be again thrown and parsing to
resume at the end of the file.

The fix is to remove the errors from the errorlist whenever a
vocabulary finishes loading successfully.

Related to #444.